Reconstruct Flask App With OOP
