+++ backend/docker-compose.yaml
@@ -10,12 +10,10 @@ services:
       - "5000:5000"
     env_file:
       - .env
-    volumes:
-      - .:/app
     expose:
       - "5000"
     healthcheck:
-      test: ["CMD", "curl", "-f", "http://localhost:$PORT"]
+      test: ["CMD-SHELL", "curl -f http://localhost:${PORT:-5000} || exit 1"]
       interval: 30s
       timeout: 10s
       retries: 3
